pairwisePenReg {pairwisePenReg}	R Documentation

high-dimensional linear regression with pairwise penalty

Description

User can choose penalty functions, including "mcp", "scad" and "lasso", where "mcp" is recommended.

Usage

  pairwisePenLinearReg(X, Y,  method='mcp', lambda=1.0, rho=1.0,lam_ridge=1e-5,
                       abstol=1e-4, reltol=1e-2, maxiter=1000,betainit=NULL, seed=1)

Arguments

X
a numeric matrix, design matrix

Y
a numeric vector, response variable

method
a string, "mcp", "scad" and "lasso" are provided

lambda
a positive real, pairwise penalty parameter

lam_ridge
a positive real, a penalty parameter for ridge penalty, a small positive real is sufficient, default as 1e-3.

Details

None.

Value

return a list, including the estimate of beta and history information of algorithm

Author(s)

Wei Liu

Examples

  dat <- gendata(n=100, p = 200, r=1, seed = 1)
  dat$beta0
  
  # estimate beta by using pairwise penalty
  resList <- pairwisePenReg(dat$X, dat$Y,  method='mcp', lambda=1.0, rho=1.0,
                              abstol=1e-4, reltol=1e-2, maxiter=1000,betainit=NULL, seed=1)
  # check the performance of estimate
  cbind(dat$beta0, resList$beta)

[Package pairwisePenReg version 1.0 Index]